Added GitHub action for compiling with GCC arm-none-eabi. Reorganized CMake configuration to allow targeting specific devices when building only STK library.

+++ test/stktest.cpp
@@ -49,7 +49,7 @@ void TestContext::ShowTestSuitePrologue()
 
 void TestContext::ShowTestSuiteEpilogue(int32_t result)
 {
-    printf("STKTEST-RESULT: %d\n", result);
+    printf("STKTEST-RESULT: %d\n", (int)result);
 }
 
 void TestContext::ForceExitTestSuie(int32_t result)
